pyproject.toml
[project]
name = "zonos"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.5.1",
    "setuptools",
    "packaging",
    "inflect>=7.5.0",
    "kanjize>=1.5.0",
    "numpy>=2.2.2",
    "phonemizer>=3.3.0",
    "sudachidict-full>=20241021",
    "sudachipy>=0.6.10",
    "torchaudio>=2.5.1",
    "transformers>=4.48.1",
    "soundfile>=0.13.1",
    "discord-py[voice]>=2.5.0",
]

# These are technically optional, but mamba-ssm is required to run hybrid models.
[project.optional-dependencies]
compile = [
    "flash-attn>=2.7.3",
    "mamba-ssm>=2.2.4",
    "causal-conv1d>=1.5.0.post8",
]

[tool.setuptools.packages.find]
include = ["zonos"]

[tool.uv]
no-build-isolation-package = ["flash-attn", "mamba-ssm", "causal-conv1d"]

[tool.ruff]
line-length = 120
